Issue #1258 - Using the MyDateFormatPropertyEditor property editor
